Refactor simulation window naming
